requirements.txt
azure-core==1.32.0
azure-storage-blob==12.24.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
cssselect2==0.7.0
defusedxml==0.7.1
idna==3.10
isodate==0.7.2
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
Pillow==9.3.0
pycparser==2.22
pygame==2.6.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
tinycss2==1.4.0
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
